Getting close
- Populated the readme;
- Created an installer;
- Removed the Bash v4 dependency;
- Restricted TTL to 5 minutes and above due to ShoutKey’s limitation.

+++ wshare.sh
@@ -5,8 +5,9 @@ set \
 	-o pipefail \
 	-o nounset
 
-shopt \
-	-s nullglob
+shopt -s \
+	nullglob \
+	expand_aliases
 
 WSHARE_HOME="$HOME/.wshare"
 
@@ -17,21 +18,21 @@ die () {
 	local message="$1"
 	local -i exit_code="${2:-1}"
 
-	echo "$message" >&2
+	echo "Error: $message" >&2
+
 	exit $exit_code
 }
 
-parse_response () {
-	local variable="$1"
-	local result="$2"
+#alias call="eval"
 
-	echo "$result" | sed "s/declare \(-[aA]\) [^ =]\{1,\}=\(.\{1,\}\)/local \1 $variable=\2/"
+result () {
+	declare -p $* | sed "s/^declare /local /"
 }
 
 upload () {
 	local file="$1"
 
-	[[ -r "$file" ]] || die "File '$file' does not exist or is not accessible" $ERR_FILE_NOT_FOUND
+	[[ -r "$file" ]] || die "File $file does not exist or is not accessible" $ERR_FILE_NOT_FOUND
 
 	local response="$(
 		cat "$file" \
@@ -44,12 +45,11 @@ upload () {
 
 	local regex="\(http://0paste.com/\([[:digit:]]\{1,\}\(-[[:alnum:]]\{1,\}\)\{0,\}\)\)\( (key: \([[:alnum:]]\{1,\}\))\)\{0,\}"
 
-	local -A result=(
-		[id]="$(echo "$response" | sed -n "s%$regex%\2%p")"
-		[url]="$(echo "$response" | sed -n "s%$regex%\1%p").txt"
-		[key]="$(echo "$response" | sed -n "s%$regex%\5%p")"
-	)
-	declare -p result
+	local upload_id="$(echo "$response" | sed -n "s%$regex%\2%p")"
+	local upload_url="$(echo "$response" | sed -n "s%$regex%\1%p").txt"
+	local upload_key="$(echo "$response" | sed -n "s%$regex%\5%p")"
+
+	result upload_id upload_url upload_key
 }
 
 shorten () {
@@ -59,23 +59,33 @@ shorten () {
 	local response="$(curl -s -d "url=$url" -d "ttl=$((ttl * 60))" "http://shoutkey.com/new")" || die "Cannot access the URL shortener service"
 
 	local regex=".*<h1>\<a href=\"\(http://shoutkey.com/[[:alpha:]]\{1,\}\)\">\([[:alpha:]]\{1,\}\)</a></h1>.*"
-	local -A result=(
-		[url]="$(echo "$response" | sed -n "s%$regex%\1%p")"
-		[key]="$(echo "$response" | sed -n "s%$regex%\2%p")"
-		[ttl]=$ttl
-	)
-	declare -p result
+
+	local shorturl_url="$(echo "$response" | sed -n "s%$regex%\1%p")"
+	local shorturl_key="$(echo "$response" | sed -n "s%$regex%\2%p")"
+	local shorturl_ttl=$ttl
+
+	result shorturl_url shorturl_key shorturl_ttl
 }
 
 delete () {
-	local file="$1"
+	local file
+	local upload_id
+	local upload_key
+	local shorturl_url
+	local response
+	for file; do
+		[[ "$file" == "$WSHARE_HOME"* ]] || continue
 
-	local upload_id="$(basename "$file")"
-	local upload_key="$(<"$file")"
+		upload_id="$(basename "$file")"
 
-	local response="$(curl -sL "http://0paste.com/$upload_id" -F "_method=delete" -F "paste[key]=$upload_key")"
+		read -r upload_key shorturl_url < "$file"
 
-	rm "$file"
+		echo "Deleting $shorturl_url at http://0paste.com/$upload_id with key $upload_key"
+
+		response="$(curl -sL "http://0paste.com/$upload_id" -F "_method=delete" -F "paste[key]=$upload_key")"
+
+		rm "$file"
+	done
 }
 
 cleanup () {
@@ -87,40 +97,41 @@ cleanup () {
 	local upload_key
 	local -a files
 	local ttl
-
+	local -i any=0
 	for ttl_dir in "$WSHARE_HOME"/*; do
 		if [[ -n "$ttl_dir" && -d "$ttl_dir" ]]; then
 			files=("$ttl_dir"/*)
 			if [[ "${#files[@]}" -gt 0 ]]; then
 				ttl="$(basename "$ttl_dir")"
 				find "$ttl_dir"/* -maxdepth 0 -type f -mmin +$ttl -exec $0 --delete {} \;
+				any=1
 			fi
 		fi
 	done
+
+	if [[ $any -eq 0 ]]; then
+		echo "No expired uploads found"
+	fi
 }
 
 upload_and_shorten () {
 	local file="$1"
-	local ttl="${2:-}"
+	local ttl="${2:-5}"
 
-	local response
-	response="$(upload "$file")"
-	if [[ -n "$response" ]]; then
-		eval "$(parse_response uploaded_file "$response")"
-		response="$(shorten "${uploaded_file[url]}" "$ttl")"
-		if [[ -n "$response" ]]; then
-			eval "$(parse_response short_url "$response")"
+	[[ $ttl -ge 5 ]] || die "TTL cannot be less than 5 minutes"
 
-			if [[ -n "${uploaded_file[key]}" ]]; then
-				local dir="$WSHARE_HOME/${short_url[ttl]}"
-				mkdir -p "$dir"
-				echo "${uploaded_file[key]}" > "$dir/${uploaded_file[id]}"
-			fi
+	eval "$(upload "$file")"
 
-			echo "${short_url[url]}"
-			echo "${short_url[url]}" | pbcopy
-		fi
+	eval "$(shorten "$upload_url" "$ttl")"
+
+	if [[ -n "$upload_key" ]]; then
+		local dir="$WSHARE_HOME/$shorturl_ttl"
+		mkdir -p "$dir"
+		echo "$upload_key $shorturl_url" > "$dir/$upload_id"
 	fi
+
+	echo "$shorturl_url"
+	echo "$shorturl_url" | pbcopy
 }
 
 show_usage () {
@@ -128,10 +139,10 @@ show_usage () {
 		Usage: $(basename $0) COMMAND
 
 		Commands:
-		    -h|--help               Shows usage
-		    -c|--clean|--cleanup    Deletes expired uploads
-		    [-s|--share] FILE [TTL] Shares a file with TTL in minutes (default = 5)
-		    -d|--delete	FILE        Deletes a file (must be under $WSHARE_HOME)
+		    -h|--help                Shows usage
+		    -c|--clean|--cleanup     Deletes expired uploads
+		    [-s|--share] FILE [TTL]  Shares a file with TTL in minutes (default: 5)
+		    -d|--delete	FILE...      Deletes files under $WSHARE_HOME (used internally)
 	EOF
 }
 
@@ -163,4 +174,39 @@ main () {
 	esac
 }
 
-main "$@"
+install () {
+	local os="$(uname)"
+	case "$os" in
+		Linux|Darwin)
+			local bin_dir="$HOME/bin"
+			mkdir -p "$bin_dir"
+
+			local bin_file="$bin_dir/wshare"
+			rm -f "$bin_file"
+			echo "$BASH_EXECUTION_STRING" > "$bin_file"
+			chmod +x "$bin_file"
+
+			if [[ "$(type -t wshare)" != "file" ]]; then
+				cat <<-EOF
+					In order to make wshare work, you need to add the following
+					to your .bashrc/.bash_profile/.profile file:
+
+					    export PATH="$HOME/bin:$PATH"
+
+				EOF
+			fi
+
+			echo "Successfully installed wshare!"
+			;;
+		*)
+			echo "Error: Unsupported OS: $os"
+			exit 1
+			;;
+	esac
+}
+
+if [[ -z "${BASH_SOURCE[0]:-}" ]]; then
+	install
+else
+	main "$@"
+fi
